cchag saved

+++ database/migrations/2023_10_14_071205_add_phone_to_users_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('mobile')->nullable()->after('email');
+            $table->bigInteger('mobile')->nullable()->after('email');
         });
     }
